Rename to orderedMap.h
